requirements.txt
albumentations==2.0.8
numpy==2.3.0
Pillow==11.2.1
scikit_learn==1.7.0
torch==2.7.1
torchvision==0.22.1
transformers==4.52.4
